Added multi variable assignment w tests

+++ tests/test.sh
@@ -120,5 +120,13 @@ if ! diff result expected15; then
   exit 1
 fi
 
+../a.out > result <<EOF
+x=y=5
+EOF
+if ! diff result expected16; then
+  echo error in test 16
+  exit 1
+fi
+
 rm result*
 echo Passed.
